confident-rag-chatbot/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
streamlit
pypdf
beautifulsoup4
requests
python-dotenv
